chore(larastan): code quality level 8.9, 3 errors remain: Line   Models/Training/Course.php  ------ --------------------------------------------------------------------------------------   139    Binary operation / between mixed and int<min, -1>|int<1, max> results in an error.          🪪  binaryOp.invalid   185    Binary operation / between mixed and int<min, -1>|int<1, max> results in an error.          🪪  binaryOp.invalid   219    Binary operation / between mixed and int<min, -1>|int<1, max> results in an error.          🪪  binaryOp.invalid

+++ app/Models/Training/Quiz/Quiz.php
@@ -74,13 +74,13 @@ class Quiz extends Model
 
     /**
      * Calculate the pass rate for this quiz
-     * @return float
+     * @return int
      */
-    public function passRate(): float
+    public function passRate(): int
     {
         $total = $this->quizResponses()->count();
         $passed = $this->quizResponses()->where('score', '>=', $this->passing_score)->count();
 
-        return $total > 0 ? ($passed / $total) * 100 : 0;
+        return $total > 0 ? (int)(($passed / $total) * 100) : 0;
     }
 }
